torii: sidecar — CustomUIHue / CustomUIAccentHue are float, not int
Caught at runtime on the next dev launch:

  [important]: [torii.ini] Failed to watch CustomUIHue: Cannot convert
    bindable of type osu.Framework.Bindables.BindableFloat retrieved
    from ConfigManager to osu.Framework.Bindables.Bindable`1[System.Int32].

The OsuConfigManager.SetDefault overload with a numeric range
(0, 359, 1) upgrades the storage to BindableFloat regardless of
whether the literal default looks int-shaped. Both hue settings use
that overload, so my type table needs typeof(float) for them, not
typeof(int).

apply-time picked the right TryParse for ints by accident (the file
held no value yet), but watchKey threw on the GetBindable<int> cast
and bailed before wiring the ValueChanged handler — so the hue
sliders silently stopped persisting back to torii.ini after a change.
Same crash mode as the original Get<object> bug, just on a different
line.

+++ osu.Game/Configuration/ToriiSettingsPersistence.cs
@@ -73,16 +73,21 @@ namespace osu.Game.Configuration
         /// </summary>
         private static readonly Dictionary<OsuSetting, Type> torii_only_settings = new Dictionary<OsuSetting, Type>
         {
-            // Custom UI hue (sesión 1 of the redesign)
+            // Custom UI hue (sesión 1 of the redesign).
+            // CustomUIHue is a BindableFloat under the hood (the
+            // SetDefault overload that takes the 0-359 range upgrades
+            // ints to float internally). Mismatched type here would
+            // throw "Cannot convert BindableFloat to Bindable<Int32>".
             { OsuSetting.CustomUIHueEnabled, typeof(bool) },
-            { OsuSetting.CustomUIHue, typeof(int) },
+            { OsuSetting.CustomUIHue, typeof(float) },
             { OsuSetting.CustomUIHueApplyToMenu, typeof(bool) },
             { OsuSetting.CustomUIHueApplyToOverlays, typeof(bool) },
             { OsuSetting.CustomUIHueApplyToSettingsPanel, typeof(bool) },
 
-            // Donator accent hue
+            // Donator accent hue (also BindableFloat for the same
+            // reason as CustomUIHue above).
             { OsuSetting.CustomUIAccentEnabled, typeof(bool) },
-            { OsuSetting.CustomUIAccentHue, typeof(int) },
+            { OsuSetting.CustomUIAccentHue, typeof(float) },
 
             // Pause/fail double-confirm (sesión 3)
             { OsuSetting.ToriiConfirmDangerousButtonsOnLongAttempts, typeof(bool) },
